docs: add support for funding sources
To support various funding sources, this commit adds a new file, `.github/FUNDING.yml`, which defines the funding model platforms. In this commit, support for the platform 'github' is added with the username 'nekofar'. Additionally, a custom funding source is added with the URL 'https://unstoppabledomains.com/d/nekofar.crypto'. This change allows users to easily contribute and support the project through multiple funding options.