ci: update build workflow to use Java 17 and restructure code inspection
This commit updates the build workflow to use Java 17 in order to take advantage of the latest language features and ensures better interoperability with current JDKs. The 'test' and 'verify' workflow jobs are now dependent on the build job for better workflow management.

Additionally, the 'inspectCode' job has been moved to run after the 'test' job, thereby making code inspection a separate process from build and test for clearer causality in case of failures. This reorganization provides a smoother CI/CD pipeline flow.