ci: update GitHub workflows to downgrade checkout version, remove java cache
The checkout action was downgraded from v3.5.3 to v3.5.2 because of known issues with the newer version.
The Java cache setting has been removed because it was causing problems with the build process. These changes were applied in both run UI tests and release workflows.
Additionally, an explicit step to set up Gradle was added using the gradle-build-action v2.6.1, to ensure the correct version of Gradle is always used.