refactor: add Vue.js support to plugin dependencies
The plugin dependency 'org.jetbrains.plugins.vue' was added to extend the functionality of the current plugin. This change will allow our plugin to properly interface with Vue.js, providing broader compatibility and utility. The dependency has been added both in the plugin.xml file and in the gradle.properties file, ensuring that it's loaded correctly during build.