ci: update release draft creation in GitHub workflow
Replaced the inline command used for creating a release draft with a more reliable GitHub action. Before, the release draft was created by a run command within the workflow which used the GitHub token. Now, a third-party GitHub action is used - nekofar/create-github-release@v1.0.8. The change was made to ease the management and improve the reliability of the release draft creation process in the workflow.